chore: rename package schemaloader to db
