+++ internal/db/pg_schema_loader.go
@@ -1,4 +1,4 @@
-package schemaloader
+package db
 
 import (
 	"context"
