+++ internal/db/schema_loader.go
@@ -1,4 +1,4 @@
-package schemaloader
+package db
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"github.com/artarts36/db-exporter/internal/schema"
 )
 
-type Loader interface {
+type SchemaLoader interface {
 	// Load database schema
 	Load(ctx context.Context) (*schema.Schema, error)
 }
